Remove double white space in script output
We had an issue where the first perfdata key of the second network interface would not display in the web interface, turns out it was the double white space in front of it causing the issue. This quick'n'dirty fix solved it.

There might be a better way to do this in the code itself.